requirements.txt
requests
web3
gnosis-py
requests_html